+++ packages/expo-screen-capture/android/src/main/java/expo/modules/screencapture/ScreenCaptureModule.kt
@@ -20,6 +20,7 @@ class ScreenCaptureModule : Module() {
   private val currentActivity
     get() = appContext.currentActivity ?: throw Exceptions.MissingActivity()
   private var screenCaptureCallback: Activity.ScreenCaptureCallback? = null
+  private var screenshotEventEmitter: ScreenshotEventEmitter? = null
 
   override fun definition() = ModuleDefinition {
     Name("ExpoScreenCapture")
@@ -33,7 +34,7 @@ class ScreenCaptureModule : Module() {
         }
         currentActivity.registerScreenCaptureCallback(currentActivity.mainExecutor, screenCaptureCallback!!)
       } else {
-        ScreenshotEventEmitter(context) {
+        screenshotEventEmitter = ScreenshotEventEmitter(context) {
           sendEvent(eventName)
         }
       }
@@ -63,7 +64,16 @@ class ScreenCaptureModule : Module() {
       currentActivity.window.clearFlags(WindowManager.LayoutParams.FLAG_SECURE)
     }.runOnQueue(Queues.MAIN)
 
+    OnActivityEntersForeground {
+      screenshotEventEmitter?.onHostResume()
+    }
+
+    OnActivityEntersBackground {
+      screenshotEventEmitter?.onHostPause()
+    }
+
     OnDestroy {
+      screenshotEventEmitter?.onHostDestroy()
       if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.UPSIDE_DOWN_CAKE) {
         screenCaptureCallback?.let {
           currentActivity.unregisterScreenCaptureCallback(it)